Move old functions to legacy.py until missing functionalities are ported
